feat: move sanitize to kafka-client.js and docstring
