+++ js/kafka/index.js
@@ -3,4 +3,3 @@ export * from "./kafka-consumer";
 export * from "./kafka-producer";
 export * from "./kafka-retry-consumer";
 export * from "./kafka-retry-producer";
-export * from "./utils";
